Inline authHeaders and executeTransactionRequest, remove ServiceHelpers
