+++ Sources/Wallet/Data/Services/Transfer/DefaultTransferService.swift
@@ -10,10 +10,32 @@ private struct TransferBody: Encodable {
     let signer: String?
 }
 
-struct DefaultTransferService: TransferService, AuthManagerBacked, TransactionExecuting {
+struct DefaultTransferService: TransferService {
     let crossmintService: CrossmintService
     let jsonCoder: JSONCoder
     let authManager: AuthManager
+
+    var authHeaders: [String: String] {
+        get async {
+            guard let jwt = await authManager.jwt else { return [:] }
+            return ["Authorization": "Bearer \(jwt)"]
+        }
+    }
+
+    func executeTransactionRequest<T: WalletTypeTransactionMapping>(
+        endpoint: Endpoint,
+        mapping: T.Type
+    ) async throws(TransactionError) -> any TransactionApiModel {
+        let data = try await crossmintService.executeRequestForRawData(
+            endpoint,
+            errorType: TransactionError.self
+        )
+        do {
+            return try jsonCoder.decode(T.APIModel.self, from: data)
+        } catch {
+            throw TransactionError.transactionGeneric("Failed to decode transaction response")
+        }
+    }
 
     func transferToken(
         _ request: TransferTokenRequest
