+++ Sources/Wallet/Data/Services/Wallet/DefaultWalletService.swift
@@ -4,12 +4,34 @@ import CrossmintService
 import Http
 import Logger
 
-struct DefaultWalletService: WalletService, AuthManagerBacked, TransactionExecuting {
+struct DefaultWalletService: WalletService {
     let crossmintService: CrossmintService
     let jsonCoder: JSONCoder
     let authManager: AuthManager
 
     var isProductionEnvironment: Bool { crossmintService.isProductionEnvironment }
+
+    var authHeaders: [String: String] {
+        get async {
+            guard let jwt = await authManager.jwt else { return [:] }
+            return ["Authorization": "Bearer \(jwt)"]
+        }
+    }
+
+    func executeTransactionRequest<T: WalletTypeTransactionMapping>(
+        endpoint: Endpoint,
+        mapping: T.Type
+    ) async throws(TransactionError) -> any TransactionApiModel {
+        let data = try await crossmintService.executeRequestForRawData(
+            endpoint,
+            errorType: TransactionError.self
+        )
+        do {
+            return try jsonCoder.decode(T.APIModel.self, from: data)
+        } catch {
+            throw TransactionError.transactionGeneric("Failed to decode transaction response")
+        }
+    }
 
     func getWallet(_ request: GetMeWalletRequest) async throws(WalletError) -> WalletApiModel {
         Logger.smartWallet.info(LogEvents.apiGetWalletStart, attributes: [
